requirements.txt
argon2-cffi==18.3.0
backports.csv==1.0.7
bcrypt==3.1.5
certifi==2018.10.15
cffi==1.11.5
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.1.2
django-crispy-forms==1.7.2
django-easy-pdf==0.1.1
django-heroku==0.3.1
django-import-export==1.2.0
django-localflavor==2.1
django-phone-field==1.7.1
et-xmlfile==1.0.1
Faker==1.0.0
gunicorn==19.9.0
html5lib==1.0.1
jdcal==1.4.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==5.3.0
psycopg2==2.8.2
pycparser==2.19
PyPDF2==1.26.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==5.1
reportlab==3.5.21
six==1.12.0
tablib==0.13.0
text-unidecode==1.2
webencodings==0.5.1
whitenoise==4.1.2
wincertstore==0.2
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
